move logging
