+++ alphadia/search_plan.py
@@ -13,11 +13,11 @@ from alphadia.outputtransform.search_plan_output import (
     SearchPlanOutput,
 )
 from alphadia.reporting import reporting
+from alphadia.reporting.logging import print_environment, print_logo
 from alphadia.search_step import (
     SearchStep,
     logger,
 )
-from alphadia.utilities.logging import print_environment, print_logo
 
 # TODO the names of the steps need to be adjusted
 TRANSFER_STEP_NAME = "transfer"
